Fixed warning.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@1188 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ include/wvbuffer.h
@@ -209,8 +209,8 @@ public:
     size_t match(const char chlist[], bool reverse = false)
         { return match((const unsigned char *)chlist, strlen(chlist),
 			 reverse); }
-    
-    int num_of_bufs()
+
+    size_t num_of_bufs()
         { return list.count(); }
 
 private:
